% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CDECGPS}
\alias{CDECGPS}
\title{CDEC gage GPS coordinates}
\format{
A data frame with 3 variables
\describe{
\item{station}{Name of the CDEC station}
\item{latitude}{latitude value in degrees}
\item{longitude}{longitude value in degree}
}
}
\usage{
CDECGPS
}
\description{
GPS Coordinates of all available CDEC stations
}
\keyword{datasets}
