% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CDECMetadata}
\alias{CDECMetadata}
\title{CDEC gage metadata}
\format{
A data frame with 7 variables
\describe{
\item{sensorDescription}{Description of the sensor}
\item{sensorNumber}{Value of the sensor}
\item{duration}{Data duration flag}
\item{plot}{Plot flag}
\item{dataCollection}{Collection source}
\item{dataAvailable}{Range of data availability}
\item{gage}{CDEC gage name}
}
}
\usage{
CDECMetadata
}
\description{
Sensor metadata of all available CDEC stations
}
\keyword{datasets}
