% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crosswalk}
\alias{crosswalk}
\title{A crosswalk of IEP survey names}
\format{
A data frame with 29 columns
\describe{
\item{variableType}{Tow duration}
\item{variable}{Maximum depth recorded at sampling location}
\item{recommendedName}{Depth range label created by the cut() function}
\item{recommendedUnit}{Length of cable to be let out}
\item{notes}{Maximum depth recorded at sampling location}
\item{ybfmp}{Depth range label created by the cut() function}
\item{YBFMP Units}{Length of cable to be let out}
\item{djfmp}{Maximum depth recorded at sampling location}
\item{DJFMP Units}{Depth range label created by the cut() function}
\item{Delta Smelt formalin preservation data}{Length of cable to be let out}
\item{Delta boat electrofishing survey }{Maximum depth recorded at sampling location}
\item{Electrofishing Units}{Depth range label created by the cut() function}
\item{EDSM}{Length of cable to be let out}
\item{EDSM Units}{Maximum depth recorded at sampling location}
\item{sls}{Depth range label created by the cut() function}
\item{SLS units}{Length of cable to be let out}
\item{20mm}{Maximum depth recorded at sampling location}
\item{20mm units}{Depth range label created by the cut() function}
\item{stn}{Length of cable to be let out}
\item{STN units}{Maximum depth recorded at sampling location}
\item{fmwt}{Depth range label created by the cut() function}
\item{FMWT units}{Length of cable to be let out}
\item{skt}{Maximum depth recorded at sampling location}
\item{SKT units}{Depth range label created by the cut() function}
\item{sms}{Length of cable to be let out}
\item{SMS units}{Maximum depth recorded at sampling location}
\item{bs}{Depth range label created by the cut() function}
}
}
\usage{
crosswalk
}
\description{
Relates IEP survey names to recommended names created by the DUWG
}
\keyword{datasets}
