% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFromITIS.R
\name{getTaxonomyFromItis}
\alias{getTaxonomyFromItis}
\title{Get full taxonomic information from ITIS}
\usage{
getTaxonomyFromItis(taxonNames, verbose = TRUE)
}
\arguments{
\item{taxonNames}{A vector of species names of their Taxonomic Species Number (TSN).
Supports different ranks as input. See details.}

\item{verbose}{Logical. Defaults to TRUE. Should progress notifications be
printed in the console?}
}
\value{
A data frame
}
\description{
Get full taxonomic information from ITIS
}
\details{
Only taxonomic rank information above the inputted value will be provided, e.g.,
if you provide a value at the phylum level, only the kingdom-associated levels
will be returned. Additionally, there may be different TSNs associated with
an input value, one being valid and the other not. Both will be returned. Users
should take care to check the 'validity' column for additional filtering
criteria.
}
\examples{
\dontrun{
taxa <- c("Chinese Mitten Crab", "Liparis", "Oncorhynchus tshawytscha")

getTaxonomyFromItis(taxa)
}
}
