% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meterSchedule}
\alias{meterSchedule}
\title{Meter schedules of IEP surveys}
\format{
A list of data frames of each survey. Each data frame has 4 columns
\describe{
\item{duration}{Tow duration}
\item{meterMin}{Minimum acceptable value}
\item{meterMax}{Maximum acceptable value}
\item{gear}{Gear type, leave as NA if none}
}
}
\usage{
meterSchedule
}
\description{
Data frame outlining the acceptable flowmeter range per tow duration
}
\keyword{datasets}
