% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdecHelpers.R
\name{pullMetadataCDEC}
\alias{pullMetadataCDEC}
\title{Pulling CDEC gage metadata}
\usage{
pullMetadataCDEC(gage, maxAttempt = 3, timeout = 60, verbose = TRUE)
}
\arguments{
\item{gage}{Name of the CDEC gage, a singular value}

\item{maxAttempt}{Number of times to retry a scrape. Defaults to 3.}

\item{timeout}{Seconds before a connection is terminated. Defaults to 60.}

\item{verbose}{Should the function annotate its progress? Defaults to TRUE.}
}
\value{
A list containing the location and sensor metadata of a CDEC station.
}
\description{
A function to pull the metadata table associated with a CDEC gage.
}
\examples{
\dontrun{
pullMetadataCDEC("MAL")
}
}
