% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullWyt.R
\name{pullWyt}
\alias{pullWyt}
\title{Scrape water year type data from DWR}
\usage{
pullWyt(url = "https://cdec.water.ca.gov/reportapp/javareports?name=WSIHIST")
}
\arguments{
\item{url}{URL of the historical Water Supply Index (WSIHIST) website}
}
\value{
A list of three tables, reconstructed wyt, eight river runoff, and official wyt classification
}
\description{
Scrape water year type data from DWR
}
\examples{
\dontrun{
pullWyt()
}
}
