% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdecHelpers.R
\name{readCdecSensorList}
\alias{readCdecSensorList}
\title{Scrape sensor list from CDEC webpage}
\usage{
readCdecSensorList(
  url = "https://cdec.water.ca.gov/reportapp/javareports?name=SensList",
  naStrings = c("N/A", "XX", "xxx"),
  naThreshold = 0.9,
  maxSearchRows = 15
)
}
\arguments{
\item{url}{Webpage to SENSLIST}

\item{naStrings}{Vector of potential NAs used in the table}

\item{naThreshold}{Threshold beyond which a column will be removed from the output}

\item{maxSearchRows}{Max number of rows to search through to find: the header row and
data start row}
}
\value{
Data frame
}
\description{
Scrape sensor list from CDEC webpage
}
\examples{
\dontrun{
readCdecSensorList()
}
}
