% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveFiles.R
\name{saveRelationalTables}
\alias{saveRelationalTables}
\title{Save the data tables pulled from an Access database as .csv files}
\usage{
saveRelationalTables(relationalTables, folderPath, fileNames = NULL)
}
\arguments{
\item{relationalTables}{A list of the relational tables, ideally taken
from \code{\link{bridgeAccess}}.}

\item{folderPath}{Path to the folder where the csv files will be saved.}

\item{fileNames}{A vector of file names (not file path). If \code{NULL}, will
pull the names from the provided \code{relationalTables} argument. Defaults to \code{NULL}}
}
\value{
Each table will be written. A metadata table will be returned
indicating success of the file writes.
}
\description{
Save the data tables pulled from an Access database as .csv files
}
\details{
This is a simple wrapper function that will save the relational tables pulled
from an Access database. The function was developed specifically for the
surveys within the Native Fish Program from CDFW, requiring the UTF-8 encoding.
}
\examples{
\dontrun{
slsTables <- bridgeAccess("https://filelib.wildlife.ca.gov/Public/Delta\%20Smelt/SLS.zip",
tables = c("Catch", "FishCodes"))

# Saving to the temporary directory
saveRelationalTables(slsTables, folderPath = tempdir())
}
}
