% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpsHelpers.R
\name{splitCoordinates}
\alias{splitCoordinates}
\title{Seperate lat/lon}
\usage{
splitCoordinates(x)
}
\arguments{
\item{x}{A vector of latitude or longitude in the format of degrees, minutes, and seconds}
}
\value{
A data frame of the coordinates in DMS
}
\description{
Seperate lat/lon
}
\keyword{internal}
