% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{towSchedule}
\alias{towSchedule}
\title{Tow schedules of IEP surveys}
\format{
A list of data frames of each survey. Each data frame has 4 columns
\describe{
\item{duration}{Tow duration}
\item{maxDepth}{Maximum depth recorded at sampling location}
\item{depth}{Depth range label created by the cut() function}
\item{cableLength}{Length of cable to be let out}
}
}
\usage{
towSchedule
}
\description{
Replicated tow schedules for IEP surveys
}
\keyword{datasets}
