% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadFiles.R
\name{getMetadataEdi}
\alias{getMetadataEdi}
\title{Grab metadata from an EDI package webpage}
\usage{
getMetadataEdi(url, version = "newest", all = FALSE)
}
\arguments{
\item{url}{URL to the EDI data package landing page}

\item{version}{Defaults to pulling the newest version. Specify a number if
you are interested in a specific version}

\item{all}{Defaults to FALSE. If TRUE, will return the XML file of the metadata
itself. If FALSE, provides only an opinionated set of parameters}
}
\value{
One of the following depending on all:
\item{list}{A list containing a data.frame of opinionated metadata, the
package title, and the publication date of the package}
\item{xml}{The EML XML file}
}
\description{
Grab metadata from an EDI package webpage
}
\details{
Only an opinionated set of metadata parameters will be returned by default,
read from the EML XML metadata file.
}
\examples{
\dontrun{
# Data package of the CDFW IEP SLS Survey
getMetadataEdi("https://portal.edirepository.org/nis/mapbrowse?packageid=edi.534.9")
}
}
