% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpsHelpers.R
\name{plotGPS}
\alias{plotGPS}
\title{A leaflet wrapper to plot station GPS coordinates with layer control.}
\usage{
plotGPS(df, layerName = "Layer", dateName = "Date", height = 1200, ...)
}
\arguments{
\item{df}{A dataframe with 6 required columns: date, station, legend, layer,
lat, and lon.}

\item{layerName}{Character vector used to label the layer element within the
pop up at each plotted point.}

\item{dateName}{Character vector used to label the date element within the
pop up at each plotted point.}

\item{height}{Height of the leaflet map.}

\item{...}{Optional. Currently only used to determine the \code{provider} argument
within the addProviderTiles.}
}
\value{
A leaflet plot.
}
\description{
A leaflet wrapper to plot station GPS coordinates with layer control.
}
\examples{
\donttest{
df <- data.frame(date = 2023,
station = c(508, 513, 520, 801),
legend = c("Theoretical"),
layer = c("Theoretical"),
lat = c(38.04717, 38.05886, 38.03217, 38.04369),
lon = c(-121.9172, -121.8677, -121.8631, -121.8440))

plotGPS(df, layerName = "Survey", dateName = "Year")
}
}
