% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdecHelpers.R
\name{pullCoordinates}
\alias{pullCoordinates}
\title{Pull CDEC gage lat/lon}
\usage{
pullCoordinates(cdecGage, maxAttempt = 3, timeout = 60)
}
\arguments{
\item{cdecGage}{Name of the gage of interest, as a character.}

\item{maxAttempt}{Number of attempts to retry the same pull}

\item{timeout}{Max duration to wait for a download, in seconds}
}
\value{
A data frame containing the station name, lat, and lon.
}
\description{
Pull CDEC gage lat/lon
}
\examples{
\dontrun{
pullCoordinates("MAL")
}
}
