% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdecHelpers.R
\name{pullMetadataCDEC}
\alias{pullMetadataCDEC}
\title{Pulling CDEC gage metadata}
\usage{
pullMetadataCDEC(station, list = T)
}
\arguments{
\item{station}{A character vector that supports one or more CDEC gage.}

\item{list}{Logical, should the output be provided as a list or data.frame?}
}
\value{
A list or data frame of the metadata associated with the requested
CDEC gage.
}
\description{
A function to pull the metadata table associated with a CDEC gage.
}
\examples{
\donttest {
pullMetadataCDEC("MAL")
}
}
