% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formScrape.R
\name{scrapeForm}
\alias{scrapeForm}
\title{Scrape, fill, and submit a static html form from the SacPas website}
\usage{
scrapeForm(url, formIndex = NULL, ..., returnForm = F)
}
\arguments{
\item{url}{URL containing the form(s) of interest}

\item{formIndex}{This argument is only required if there are more than 1 form
on the webpage. A numeric index indicating the form of interest}

\item{...}{Field names with their appropriate options to submit as a query.
Fields allowing multiple selections should be provided as a vector.}

\item{returnForm}{Logical, will return the html form if TRUE}
}
\value{
An \code{rvest_session}
}
\description{
Scrape, fill, and submit a static html form from the SacPas website
}
\examples{
\dontrun{
# This specific query has duplicated field names across the checkboxes and a
# select field. The function creates a generic label to make these unique,
# however, these are uninformative labels. To know which ones you want,
# you can look at the website or the form itself. Here, the latter works

scrapeForm("https://www.cbr.washington.edu/sacramento/data/query_river_graph.html",
returnForm = T)

# we want "WaterTempAvg" as the 10 year avg value, which is `data[]_7`:

scrapeForm("https://www.cbr.washington.edu/sacramento/data/query_river_graph.html",
hafilter = "All", `year[]` = 2023, outputFormat = "graph", tempUnit = "C",
`loc[]` = c("RDB", "WLK"), `data[]` = "WaterTemperature",
avgyear = 2023, size = "large", `data[]_2` = F, `data[]_3` = F,
`data[]_4` = F, `data[]_5` = F, `data[]_6` = F,
monochrome = F, datalink = F)
}
}
